update copyright, compare strtol() output to LONG_MAX, not INT_MAX

+++ params.c
@@ -1,3 +1,26 @@
+/* $Id: params.c,v 1.3 2007/01/17 20:55:01 aaron Exp $ */
+/* Copyright 2006-2007 Codemass, Inc.  All rights reserved.
+ * Use is subject to license terms.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * @file params.c
+ * @brief Input parameter parsing and help display implementation.
+ * @author Aaron Bannert (aaron@codemass.com)
+ */
+
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
@@ -109,7 +132,7 @@ void parse_args(int argc, char *argv[])
                     perror("invalid concurrency (-c) value");
                     print_help(stderr, progname);
                     exit(-1);
-                } else if (l <= 0 || l == INT_MAX) {
+                } else if (l <= 0 || l == LONG_MAX) {
                     fprintf(stderr, "invalid concurrency (-c): %ld\n", l);
                     print_help(stderr, progname);
                     exit(-1);
@@ -123,7 +146,7 @@ void parse_args(int argc, char *argv[])
                     perror("invalid request count (-n) value");
                     print_help(stderr, progname);
                     exit(-1);
-                } else if (l <= 0 || l == INT_MAX) {
+                } else if (l <= 0 || l == LONG_MAX) {
                     fprintf(stderr, "invalid request count (-n): %ld\n", l);
                     print_help(stderr, progname);
                     exit(-1);
